Restore obsolete plugins to functionality for now
The plugins restored here are largely obsolete, preferring to use
jruby:exec and similar targets to run normal external Ruby
programs, but they are still dependencies of other projects that
have not migrated (such as jruby-rack). This commit and the two
prior restore the plugins and update them in the same way the ones
were upated to recent JRuby and gem versions.